day15/2 solved

+++ app/src/main/java/com/vimacodes/aoc/day15/HASHAlgorithm.java
@@ -1,8 +1,8 @@
 package com.vimacodes.aoc.day15;
 
 class HASHAlgorithm {
-  public static long hash(String input) {
-    long result = 0;
+  public static int hash(String input) {
+    int result = 0;
 
     for (int i = 0; i < input.length(); i++) {
       result += input.codePointAt(i);
